# docs/reference/api/openapi.yaml
openapi: 3.0.3
info:
  title: ICN Gateway API
  description: |
    REST API for the Intercooperative Network (ICN) Gateway.
    
    ICN is a substrate daemon for the cooperative internet, providing:
    - Decentralized identity (DIDs)
    - Trust graph computation
    - Mutual credit ledger
    - Cooperative contract execution
    - Distributed compute coordination
    - Democratic governance primitives
  version: 0.1.0
  contact:
    name: ICN Development Team
    url: https://github.com/InterCooperative-Network/icn
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://gateway.icn.example
    description: Production gateway
...
security:
  - BearerAuth: []
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      description: Get notifications for the authenticated user
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
  /notifications/{id}/read:
    post:
      tags: [Notifications]
      summary: Mark as read
      description: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        created_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true