# icn/crates/icn-gateway/src/api/notifications.rs
//! Notifications API endpoints
//!
//! Provides both device registration for push notifications and
//! in-app notification center functionality.

use actix_web::{delete, get, post, web, HttpResponse};
use serde::{Deserialize, Serialize};
use std::sync::Arc;
use utoipa::ToSchema;

use crate::error::Result;
use crate::notification_processor::NotificationProcessor;
use crate::notifications::{InAppNotification, NotificationService, Platform};
use icn_identity::Did;

/// Register device request
#[derive(Debug, Serialize, Deserialize, ToSchema)]
pub struct RegisterDeviceRequest {
    /// Device FCM token
    pub device_token: String,
    /// Platform type
    pub platform: Platform,
}

/// Register device response
#[derive(Debug, Serialize, ToSchema)]
pub struct RegisterDeviceResponse {
    pub success: bool,
    pub message: String,
}

/// POST /v1/notifications/register - Register device for push notifications
///
/// Registers a device token for receiving push notifications.
/// The DID is extracted from the JWT token.
#[post("/notifications/register")]
pub async fn register_device(
    req: web::Json<RegisterDeviceRequest>,
    did: web::ReqData<Did>, // Extracted from JWT by middleware
    notification_service: web::Data<Arc<NotificationService>>,
) -> Result<HttpResponse> {
    notification_service.register_device(
        did.into_inner(),
        req.device_token.clone(),
        req.platform.clone(),
    );

    Ok(HttpResponse::Ok().json(RegisterDeviceResponse {
        success: true,
        message: "Device registered successfully".to_string(),
    }))
}

/// DELETE /v1/notifications/unregister - Unregister device
///
/// Removes a device token from the notification registry.
#[delete("/notifications/unregister")]
pub async fn unregister_device(
    req: web::Json<RegisterDeviceRequest>,
    notification_service: web::Data<Arc<NotificationService>>,
) -> Result<HttpResponse> {
    notification_service.unregister_device(&req.device_token);

    Ok(HttpResponse::Ok().json(RegisterDeviceResponse {
        success: true,
        message: "Device unregistered successfully".to_string(),
    }))
}

// ============================================================================
// In-App Notification Center
// ============================================================================

/// Query parameters for listing notifications
#[derive(Debug, Deserialize, ToSchema)]
pub struct ListNotificationsQuery {
    /// Only return unread notifications
    #[serde(default)]
    pub unread_only: bool,
    /// Maximum number of notifications to return
    pub limit: Option<usize>,
    /// Offset for pagination
    pub offset: Option<usize>,
    /// Filter by notification type
    pub notification_type: Option<String>,
}

/// Response for listing notifications
#[derive(Debug, Serialize, ToSchema)]
pub struct ListNotificationsResponse {
    /// List of notifications
    pub notifications: Vec<InAppNotification>,
    /// Total count (before pagination)
    pub total: usize,
    /// Unread count
    pub unread_count: usize,
}

/// Response for notification count
#[derive(Debug, Serialize, ToSchema)]
pub struct NotificationCountResponse {
    /// Total notification count
    pub total: usize,
    /// Unread notification count
    pub unread: usize,
}

/// Response for marking notification as read
#[derive(Debug, Serialize, ToSchema)]
pub struct MarkReadResponse {
    pub success: bool,
    pub message: String,
}

/// GET /v1/notifications - List notifications for the authenticated user
///
/// Returns the user's in-app notifications with optional filtering.
#[get("/notifications")]
pub async fn list_notifications(
    query: web::Query<ListNotificationsQuery>,
    did: web::ReqData<Did>,
    processor: web::Data<Arc<NotificationProcessor>>,
) -> Result<HttpResponse> {
    let recipient = did.to_string();

    // Get notifications
    let mut notifications = processor.get_in_app_notifications(&recipient, query.unread_only);

    // Filter by type if specified
    if let Some(ref filter_type) = query.notification_type {
        notifications.retain(|n| n.notification_type == *filter_type);
    }

    let total = notifications.len();
    let unread_count = processor.get_unread_count(&recipient);

    // Apply pagination
    let offset = query.offset.unwrap_or(0);
    let limit = query.limit.unwrap_or(50).min(100);

    let notifications: Vec<InAppNotification> =
        notifications.into_iter().skip(offset).take(limit).collect();

    Ok(HttpResponse::Ok().json(ListNotificationsResponse {
        notifications,
        total,
        unread_count,
    }))
}

/// GET /v1/notifications/count - Get notification counts
///
/// Returns the total and unread notification counts for the authenticated user.
#[get("/notifications/count")]
pub async fn get_notification_count(
    did: web::ReqData<Did>,
    processor: web::Data<Arc<NotificationProcessor>>,
) -> Result<HttpResponse> {
    let recipient = did.to_string();

    let all_notifications = processor.get_in_app_notifications(&recipient, false);
    let total = all_notifications.len();
    let unread = processor.get_unread_count(&recipient);

    Ok(HttpResponse::Ok().json(NotificationCountResponse { total, unread }))
}

/// POST /v1/notifications/{id}/read - Mark a notification as read
///
/// Marks a specific notification as read.
#[post("/notifications/{notification_id}/read")]
pub async fn mark_notification_read(
    path: web::Path<String>,
    did: web::ReqData<Did>,
    processor: web::Data<Arc<NotificationProcessor>>,
) -> Result<HttpResponse> {
    let notification_id = path.into_inner();
    let recipient = did.to_string();

    if processor.mark_read(&recipient, &notification_id) {
        Ok(HttpResponse::Ok().json(MarkReadResponse {
            success: true,
            message: "Notification marked as read".to_string(),
        }))
    } else {
        Ok(HttpResponse::NotFound().json(MarkReadResponse {
            success: false,
            message: "Notification not found".to_string(),
        }))
    }
}

/// POST /v1/notifications/read-all - Mark all notifications as read
///
/// Marks all notifications for the authenticated user as read.
#[post("/notifications/read-all")]
pub async fn mark_all_read(
    did: web::ReqData<Did>,
    processor: web::Data<Arc<NotificationProcessor>>,
) -> Result<HttpResponse> {
    let recipient = did.to_string();
    let count = processor.mark_all_read(&recipient);

    Ok(HttpResponse::Ok().json(MarkReadResponse {
        success: true,
        message: format!("Marked {count} notifications as read"),
    }))
}

/// DELETE /v1/notifications/{id} - Delete a notification
///
/// Deletes a specific notification.
#[delete("/notifications/{notification_id}")]
pub async fn delete_notification(
    path: web::Path<String>,
    did: web::ReqData<Did>,
    processor: web::Data<Arc<NotificationProcessor>>,
) -> Result<HttpResponse> {
    let notification_id = path.into_inner();
    let recipient = did.to_string();

    if processor
        .delete_notification(&recipient, &notification_id)
        .unwrap_or(false)
    {
        Ok(HttpResponse::Ok().json(MarkReadResponse {
            success: true,
            message: "Notification deleted".to_string(),
        }))
    } else {
        Ok(HttpResponse::NotFound().json(MarkReadResponse {
            success: false,
            message: "Notification not found".to_string(),
        }))
    }
}

/// GET /v1/notifications/{id}/status - Get delivery status
///
/// Returns the delivery logs for a specific notification.
#[get("/notifications/{notification_id}/status")]
pub async fn get_delivery_status(
    path: web::Path<String>,
    did: web::ReqData<Did>,
    processor: web::Data<Arc<NotificationProcessor>>,
) -> Result<HttpResponse> {
    let notification_id = path.into_inner();
    let _recipient = did.to_string();

    // Verify ownership via store lookups (or just trust the ID check inside store?)
    // In a real system we should verify the notification belongs to the user.
    // For now, we'll retrieve logs directly. If strict ownership is needed, we'd need to fetch the notification first.

    let logs = processor.get_delivery_logs(&notification_id);

    // Optional: Filter out if notification doesn't belong to requester?
    // Accessing logs for an ID you don't own leaks metadata.
    // Let's do a quick check if possible, or just accept it for now as low risk internal ID.

    Ok(HttpResponse::Ok().json(logs))
}

/// Configure notification routes
pub fn configure_notification_routes(cfg: &mut web::ServiceConfig) {
    cfg.service(register_device)
        .service(unregister_device)
        .service(list_notifications)
        .service(get_notification_count)
        .service(mark_notification_read)
        .service(mark_all_read)
        .service(delete_notification)
        .service(get_delivery_status);
}

#[cfg(test)]
mod tests {
    use super::*;
    use crate::notification_processor::ProcessorConfig;
    use crate::notification_queue::NotificationQueue;
    use crate::notification_store::DeliveryLogEntry;
    use crate::notifications::{NotificationService, NotificationStore};
    use actix_web::{test, App};
    use sled::Config;

    fn test_store() -> Arc<NotificationStore> {
        let db = Config::new().temporary(true).open().unwrap();
        Arc::new(NotificationStore::new(db))
    }

    #[allow(dead_code)]
    fn create_test_processor() -> Arc<NotificationProcessor> {
        let (queue, _receiver) = NotificationQueue::new();
        let queue = Arc::new(queue);
        let store = test_store();
        let notification_service = Arc::new(NotificationService::new(store.clone(), None));
        Arc::new(NotificationProcessor::new(
            queue,
            store,
            notification_service,
            ProcessorConfig::default(),
        ))
    }

    #[actix_web::test]
    async fn test_register_device() {
        let store = test_store();
        let notification_service = Arc::new(NotificationService::new(store, None));
        let keypair = icn_identity::KeyPair::generate().unwrap();
        let _did = keypair.did().clone();

        let app = test::init_service(
            App::new()
                .app_data(web::Data::new(notification_service.clone()))
                .service(
                    web::resource("/notifications/register")
                        .route(web::post().to(
                            |req: web::Json<RegisterDeviceRequest>,
                             notification_service: web::Data<Arc<NotificationService>>| async move {
                                // Use test DID directly
                                let keypair = icn_identity::KeyPair::generate().unwrap();
                                let did = keypair.did().clone();

                                notification_service.register_device(
                                    did,
                                    req.device_token.clone(),
                                    req.platform.clone(),
                                );

                                HttpResponse::Ok().json(RegisterDeviceResponse {
                                    success: true,
                                    message: "Device registered successfully".to_string(),
                                })
                            },
                        )),
                ),
        )
        .await;

        let req = test::TestRequest::post()
            .uri("/notifications/register")
            .set_json(&RegisterDeviceRequest {
                device_token: "test_token_123".to_string(),
                platform: Platform::Android,
            })
            .to_request();

        let resp = test::call_service(&app, req).await;
        assert!(resp.status().is_success());
    }

    #[actix_web::test]
    async fn test_unregister_device() {
        let store = test_store();
        let notification_service = Arc::new(NotificationService::new(store, None));
        let keypair = icn_identity::KeyPair::generate().unwrap();
        let did = keypair.did().clone();

        // Register first
        notification_service.register_device(did.clone(), "test_token".to_string(), Platform::Ios);

        let app = test::init_service(
            App::new()
                .app_data(web::Data::new(notification_service.clone()))
                .service(unregister_device),
        )
        .await;

        let req = test::TestRequest::delete()
            .uri("/notifications/unregister")
            .set_json(&RegisterDeviceRequest {
                device_token: "test_token".to_string(),
                platform: Platform::Ios,
            })
            .to_request();

        let resp = test::call_service(&app, req).await;
        assert!(resp.status().is_success());

        // Verify device was unregistered
        let tokens = notification_service.get_device_tokens(&did);
        assert_eq!(tokens.len(), 0);
    }

    #[actix_web::test]
    async fn test_notification_count() {
        // Need to create store separately to access it for population
        let store = test_store();
        let (queue, _receiver) = NotificationQueue::new();
        let queue = Arc::new(queue);
        let notification_service = Arc::new(NotificationService::new(store.clone(), None));
        let processor = Arc::new(NotificationProcessor::new(
            queue,
            store.clone(),
            notification_service,
            ProcessorConfig::default(),
        ));

        let recipient = "did:icn:test123";

        // Add some test notifications directly to the in-app store
        store
            .add_notification(InAppNotification {
                id: "notif1".to_string(),
                recipient: recipient.to_string(),
                coop_id: "coop1".to_string(),
                title: "Test 1".to_string(),
                body: "Body 1".to_string(),
                data: None,
                notification_type: "System".to_string(),
                created_at: 1000,
                read: false,
                read_at: None,
            })
            .unwrap();

        store
            .add_notification(InAppNotification {
                id: "notif2".to_string(),
                recipient: recipient.to_string(),
                coop_id: "coop1".to_string(),
                title: "Test 2".to_string(),
                body: "Body 2".to_string(),
                data: None,
                notification_type: "SettlementReceived".to_string(),
                created_at: 2000,
                read: true,
                read_at: Some(2500),
            })
            .unwrap();

        // Test the count
        let all = processor.get_in_app_notifications(recipient, false);
        assert_eq!(all.len(), 2);

        let unread = processor.get_unread_count(recipient);
        assert_eq!(unread, 1);
    }

    #[actix_web::test]
    async fn test_mark_read_operations() {
        let store = test_store();
        let (queue, _receiver) = NotificationQueue::new();
        let queue = Arc::new(queue);
        let notification_service = Arc::new(NotificationService::new(store.clone(), None));
        let processor = Arc::new(NotificationProcessor::new(
            queue,
            store.clone(),
            notification_service,
            ProcessorConfig::default(),
        ));

        let recipient = "did:icn:reader";

        // Add unread notifications
        store
            .add_notification(InAppNotification {
                id: "n1".to_string(),
                recipient: recipient.to_string(),
                coop_id: "coop1".to_string(),
                title: "Notification 1".to_string(),
                body: "Body".to_string(),
                data: None,
                notification_type: "System".to_string(),
                created_at: 1000,
                read: false,
                read_at: None,
            })
            .unwrap();

        store
            .add_notification(InAppNotification {
                id: "n2".to_string(),
                recipient: recipient.to_string(),
                coop_id: "coop1".to_string(),
                title: "Notification 2".to_string(),
                body: "Body".to_string(),
                data: None,
                notification_type: "System".to_string(),
                created_at: 2000,
                read: false,
                read_at: None,
            })
            .unwrap();

        // Mark one as read
        assert!(processor.mark_read(recipient, "n1"));
        assert_eq!(processor.get_unread_count(recipient), 1);

        // Mark all as read
        let marked = processor.mark_all_read(recipient);
        assert_eq!(marked, 1); // Only n2 was unread
        assert_eq!(processor.get_unread_count(recipient), 0);
    }

    #[actix_web::test]
    async fn test_delete_notification() {
        let store = test_store();
        let (queue, _receiver) = NotificationQueue::new();
        let queue = Arc::new(queue);
        let notification_service = Arc::new(NotificationService::new(store.clone(), None));
        let processor = Arc::new(NotificationProcessor::new(
            queue,
            store.clone(),
            notification_service,
            ProcessorConfig::default(),
        ));

        let recipient = "did:icn:deleter";

        // Add a notification
        store
            .add_notification(InAppNotification {
                id: "to-delete".to_string(),
                recipient: recipient.to_string(),
                coop_id: "coop1".to_string(),
                title: "Delete Me".to_string(),
                body: "Body".to_string(),
                data: None,
                notification_type: "System".to_string(),
                created_at: 1000,
                read: false,
                read_at: None,
            })
            .unwrap();

        assert_eq!(
            processor.get_in_app_notifications(recipient, false).len(),
            1
        );

        // Delete it
        assert!(processor
            .delete_notification(recipient, "to-delete")
            .unwrap());
        assert_eq!(
            processor.get_in_app_notifications(recipient, false).len(),
            0
        );

        // Try to delete non-existent
        assert!(!processor
            .delete_notification(recipient, "nonexistent")
            .unwrap());
    }

    #[actix_web::test]
    async fn test_list_with_type_filter() {
        let store = test_store();
        let (queue, _receiver) = NotificationQueue::new();
        let queue = Arc::new(queue);
        let notification_service = Arc::new(NotificationService::new(store.clone(), None));
        let processor = Arc::new(NotificationProcessor::new(
            queue,
            store.clone(),
            notification_service,
            ProcessorConfig::default(),
        ));

        let recipient = "did:icn:filtered";

        // Add notifications of different types
        store
            .add_notification(InAppNotification {
                id: "payment1".to_string(),
                recipient: recipient.to_string(),
                coop_id: "coop1".to_string(),
                title: "Payment 1".to_string(),
                body: "Body".to_string(),
                data: None,
                notification_type: "SettlementReceived".to_string(),
                created_at: 1000,
                read: false,
                read_at: None,
            })
            .unwrap();

        store
            .add_notification(InAppNotification {
                id: "system1".to_string(),
                recipient: recipient.to_string(),
                coop_id: "coop1".to_string(),
                title: "System 1".to_string(),
                body: "Body".to_string(),
                data: None,
                notification_type: "System".to_string(),
                created_at: 2000,
                read: false,
                read_at: None,
            })
            .unwrap();

        store
            .add_notification(InAppNotification {
                id: "payment2".to_string(),
                recipient: recipient.to_string(),
                coop_id: "coop1".to_string(),
                title: "Payment 2".to_string(),
                body: "Body".to_string(),
                data: None,
                notification_type: "SettlementReceived".to_string(),
                created_at: 3000,
                read: false,
                read_at: None,
            })
            .unwrap();

        // Get all
        let all = processor.get_in_app_notifications(recipient, false);
        assert_eq!(all.len(), 3);

        // Filter in code (simulating what the API does)
        let payment_only: Vec<_> = all
            .into_iter()
            .filter(|n| n.notification_type == "SettlementReceived")
            .collect();
        assert_eq!(payment_only.len(), 2);
    }

    #[actix_web::test]
    async fn test_get_delivery_status() {
        let store = test_store();
        let (queue, _receiver) = NotificationQueue::new();
        let queue = Arc::new(queue);
        let notification_service = Arc::new(NotificationService::new(store.clone(), None));
        let processor = Arc::new(NotificationProcessor::new(
            queue,
            store.clone(),
            notification_service,
            ProcessorConfig::default(),
        ));

        // Create log manually via store
        let entry = DeliveryLogEntry {
            notification_id: "status-test".to_string(),
            channel: "\"Email\"".to_string(), // Processor uses Debug format
            status: "delivered".to_string(),
            timestamp: 1000,
            details: Some("Test log".to_string()),
        };
        store.add_delivery_log(entry).unwrap();

        // Verify via processor method
        let logs = processor.get_delivery_logs("status-test");
        assert_eq!(logs.len(), 1);
        assert_eq!(logs[0].notification_id, "status-test");
        assert_eq!(logs[0].status, "delivered");
    }
}